W435: 3 functions matched, first-party nd -327, zero regressions
First-party MATCH 6,888 -> 6,891 of 8,235 (83.7%).
Total MATCH 8,054 -> 8,057. NONMATCHING 1,343.
12 functions improved, 0 worsened. Zero MATCHes lost, zero blocking statuses,
zero new over-window functions.

Newly matching
  002d2c10  Battle/btlTarget.c            nd 10 -> 0
  003182d0  Graphics/Model/mdlManager.c   nd  2 -> 0  mdlAnimSet
  003b7090  Scene/mt_scene.c              nd  2 -> 0

Largest reductions
  003d8080  Fcl/fclCombineMisc.c    173 ->  14
  00320080  Graphics/Model/mdlFile.c 167 ->  51
  001fba70  Battle/Result/br_data.c  239 -> 230
  001fbfa0  Battle/Result/br_data.c  118 -> 110

The previous two waves removed 4,481 nd and converted nothing, so this one
targeted conversions directly. The pool was every function whose prologue
already matches retail exactly, whose object equals retail's real code length,
and whose nd is low. That combination closes the diagnosis space: nothing is
missing, nothing is surplus, the instruction count is right, so the residual can
only be arrangement - register role, commutative operand order, statement
placement, global addressing mode, jump-table data, or narrow width. Seventy-one
functions qualified initially, and lanes widened it to roughly two hundred.

Two rules were corrected mid-wave and both mattered.

The pool was first defined as object_size == window, which is wrong: retail pads
functions to a 16-byte boundary, so a function ending four or eight bytes short
still emits retail's exact instruction count. The correct comparison is against
retail's REAL code length - the window minus its trailing run of zero or nop
words. That widening roughly tripled the pool.

Relatedly, "the object must not change size" was too crude. The real rule is
that a change is acceptable if the object lands at or below retail's real code
length while vacating only padding, and is rejected only if it exceeds the
window or vacates a real instruction. mdlFile 00320080 was rejected under the
old rule and is worth -116 under the new one; fclCombineMisc 003d8080 was
rejected the same way and is worth -159.

One technique overturned a standing assumption. A commutative addu whose
operands include a base ADDRESS is not an arithmetic-order problem. Four inline
helper variants on mdlAnimSet - u32 and pointer typed, both parameter orders -
all made it worse across two waves. What converted it was declaring a pointer
local at the TAIL of the block and folding its assignment into the expression
that uses it, so the address is materialised at the point of use. Tail placement
of the new declaration has now decided four separate conversions.

mt_scene 003b7090 closed on a missing call argument, and btlTarget 002d2c10 on a
tail-declared loop index - both cause-1 liveness rather than colouring.

A dedicated audit lane ran twenty-eight fresh verify sweeps over the pool
looking for the failure that most often costs a banked conversion: a function at
nd 0 whose marker still carries the trailing NONMATCHING and therefore reports
STALE_NONMATCHING instead of MATCH. It found no current-state instances and
confirmed all three conversions banked correctly.

Verified: tools/build.py reproduces SLUS_216.21 sha1
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb; 92 tests pass.

+++ src/Graphics/Model/mdlFile.c
@@ -5641,7 +5641,7 @@ void FUN_00320080(int param_4,float param_1,int param_5,float param_2,float para
 
   if (piVar2 == *(int **)(param_5 + 0x50)) {
 
-    u16 uVar1 = *(u16 *)(piVar2 + 1);
+    u32 uVar1 = *(u16 *)(piVar2 + 1);
 
     for (uVar6 = 0; uVar6 < 4; uVar6 = uVar6 + 1) {
